cleaned up some files
