+++ src/api/entries.json
@@ -20,7 +20,20 @@
       "concepts": "booty",
       "entry": "Today......",
       "mood": "😀"
+    },
+    {
+      "date": "2020-06-09",
+      "concepts": "adsf",
+      "entry": "A penny for your thoughadfadsfts...",
+      "mood": "🌟🕺🌟",
+      "id": 4
+    },
+    {
+      "date": "2020-06-16",
+      "concepts": "wowowow",
+      "entry": "fook off",
+      "mood": "🤬",
+      "id": 5
     }
-    
   ]
 }
