+++ src/scripts/entryGrabber.js
@@ -1,18 +1,18 @@
-import API from "./data.js"
+// import API from "./data.js"
 
-// grabs new entry data 
+// // grabs new entry data 
 
 
 
-document.getElementById("#myJournalButton").addEventListener("click", event => {
+// document.querySelector("#myJournalButton").addEventListener("click", event => {
 
-    const journalEntryObj = {}
+//     const journalEntryObj = {}
 
-    journalEntryObj.date = document.querySelector("#journalDate").value
-    journalEntryObj.concepts = document.querySelector("#journalConcepts").value
-    journalEntryObj.entry = document.querySelector("#journalEntry").value
-    journalEntryObj.mood = document.querySelector("#mood").value
+//     journalEntryObj.date = document.querySelector("#journalDate").value
+//     journalEntryObj.concepts = document.querySelector("#journalConcepts").value
+//     journalEntryObj.entry = document.querySelector("#journalEntry").value
+//     journalEntryObj.mood = document.querySelector("#mood").value
 
-    
-    API.saveJournalEntry(journalEntryObj)
-})
+//     console.log(journalEntryObj)
+//     API.saveJournalEntry(journalEntryObj)
+// })
