--- src/scripts/main.js
@@ -1,3 +1,0 @@
-// API.getJournalData()
-//     .then((entryArray) => {journalEntries = entryArray})
-//     .then(() => {journalPost()})
