--- src/scripts/saveJournalEntry.js
@@ -1,15 +1,0 @@
-// saves api to local json database
-
-
-// const API = {
-//     saveJournalEntry: (newEntryObject) => {
-//         return fetch("http://localhost:8088/entries", {
-//             method: "POST",
-//             headers: {
-//                 "Content-Type": "application/json"
-//             },
-//             body: JSON.stringify(newEntryObject)
-//         })
-//     }
-// }
-
